Created AlertPresenter.
