+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,15 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Александра Коснырева on 07.09.2023.
+//
+
+import Foundation
+
+
+ struct QuizQuestion {
+     let image: String
+     let text: String
+     let correctAnswer: Bool
+ }
