+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Александра Коснырева on 07.09.2023.
+//
+
+import Foundation
+import UIKit
+
+
+struct QuizStepViewModel{
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
