--- MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -1,182 +1,0 @@
-import UIKit
-
-final class MovieQuizViewController: UIViewController {
-    // MARK: - Lifecycle
-    @IBOutlet weak var noButton: UIButton!
-    
-    @IBOutlet weak var yesButton: UIButton!
-    
-    @IBOutlet weak var questionTitleLabel: UILabel!
-    
-    @IBOutlet private var imageView: UIImageView!
-    
-    @IBOutlet private var questionLabel: UILabel!
-    
-    @IBOutlet private var indexLabel: UILabel!
-    
-    struct QuizStepViewModel{
-        let image: UIImage
-        let question: String
-        let questionNumber: String
-    }
-    struct QuizResultsViewModel{
-        let title: String
-        let text: String
-        let buttonText: String
-    }
-    struct QuizQuestion {
-        let image: String
-        let text: String
-        let correctAnswer: Bool
-    }
-    
-    private let questions: [QuizQuestion] = [
-        QuizQuestion(image: "The Godfather", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-        QuizQuestion(image: "The Dark Knight", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-        QuizQuestion(image: "Kill Bill", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-        QuizQuestion(image: "The Avengers", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-        QuizQuestion(image: "Deadpool", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-        QuizQuestion(image: "The Green Knight", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: true),
-        QuizQuestion(image: "Old", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false),
-        QuizQuestion(image: "The Ice Age Adventures of Buck Wild", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false),
-        QuizQuestion(image: "Tesla", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false),
-        QuizQuestion(image: "Vivarium", text: "Рейтинг этого фильма больше чем 6?", correctAnswer: false)
-    ]
-    
-    private var currentQuestionIndex = 0 //текущий индекс
-    private var correctAnswers = 0
-    
-    private func convert(model: QuizQuestion) -> QuizStepViewModel{
-        let questionStep = QuizStepViewModel(image: UIImage(named: model.image) ?? UIImage(), question: model.text, questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
-        
-        return questionStep
-    }
-    private func show (quiz step: QuizStepViewModel) {
-        indexLabel.text = step.questionNumber
-        imageView.image = step.image
-        questionLabel.text = step.question
-        imageView.layer.borderColor = UIColor.clear.cgColor
-    }
-    private func showAnswerResult(isCorrect: Bool) {
-        if isCorrect {
-            correctAnswers += 1
-        }
-        //метод красит рамку
-        imageView.layer.masksToBounds = true
-        imageView.layer.borderWidth = 8
-        imageView.layer.borderColor = isCorrect ? UIColor.ypGreen.cgColor : UIColor.ypRed.cgColor
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
-            //код который должен быть вызван через одну секунду
-            self.showNextQuestionOrResults()
-        }
-    }
-    private func showNextQuestionOrResults(){
-        if currentQuestionIndex == questions.count - 1 {// идем в состояние результата квиза
-            let text = "Ваш результат: \(correctAnswers)/10"
-            let viewModel = QuizResultsViewModel(title: "Этот раунд окончен!", text: text, buttonText: "Сыграть еще раз")
-            show(quiz: viewModel)
-        } else { // показываем следующий вопрос
-            currentQuestionIndex += 1
-            let nextQuestion = questions[currentQuestionIndex]
-            let viewModel = convert(model: nextQuestion)
-            show(quiz: viewModel)
-        }
-    }
-    //Когда заканчивается квиз появляется системный алерт, содержащий инфу о пройденном квизе и предлогающий запустить квиз еще раз
-    // создадим кнопку для запуска (константа с кнопкой для системного алерта)
-    private func show(quiz result: QuizResultsViewModel) {
-        let alert = UIAlertController(title: "Этот раунд окончен!", message: "Ваш результат \(correctAnswers)", preferredStyle: .alert)
-        let action = UIAlertAction(title: "Сыграть ещё раз", style: .default) { _ in
-            // код, который сбрасывает игру и показывает первый вопрос
-            self.currentQuestionIndex = 0
-            self.correctAnswers = 0
-            let firstQuestion = self.questions[self.currentQuestionIndex]
-            let viewModel = self.convert(model: firstQuestion)
-            self.show(quiz: viewModel)
-        }
-            alert.addAction(action)
-            self.present(alert, animated: true, completion: nil)
-        }
-    //метод который вызывается когда пользователь жмет на кнопку нет
-    @IBAction private func noButtonClicked(_ sender: UIButton) {
-        let currentQuestion = questions[currentQuestionIndex]
-        let givenResult = false
-        showAnswerResult(isCorrect: givenResult == currentQuestion.correctAnswer)
-    }
-    //метод когда пользователь жмет на кнопку да
-    @IBAction private func yesButtonClicked(_ sender: UIButton) {
-        let currentQuestion = questions[currentQuestionIndex]
-        let givenResult = true
-        showAnswerResult(isCorrect: givenResult == currentQuestion.correctAnswer)
-    }
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        yesButton.titleLabel!.font = UIFont(name: "YSDisplay-Medium", size: 20)
-        noButton.titleLabel!.font = UIFont(name: "YSDisplay-Medium", size: 20)
-        questionLabel.font = UIFont(name: "YSDisplay-Bold", size: 23)
-        indexLabel.font = UIFont(name: "YSDisplay-Medium", size: 20)
-    }
-}
-/*
- Mock-данные
- 
- 
- Картинка: The Godfather
- Настоящий рейтинг: 9,2
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
-
-
- Картинка: The Dark Knight
- Настоящий рейтинг: 9
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
-
-
- Картинка: Kill Bill
- Настоящий рейтинг: 8,1
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
-
-
- Картинка: The Avengers
- Настоящий рейтинг: 8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
-
-
- Картинка: Deadpool
- Настоящий рейтинг: 8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
-
-
- Картинка: The Green Knight
- Настоящий рейтинг: 6,6
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
-
-
- Картинка: Old
- Настоящий рейтинг: 5,8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
-
-
- Картинка: The Ice Age Adventures of Buck Wild
- Настоящий рейтинг: 4,3
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
-
-
- Картинка: Tesla
- Настоящий рейтинг: 5,1
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
-
-
- Картинка: Vivarium
- Настоящий рейтинг: 5,8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
- */
