+++ MovieQuiz/Services/AlertPresenerProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  AlertPresenerProtocol.swift
+//  MovieQuiz
+//
+//  Created by Александра Коснырева on 07.09.2023.
+//
+
+import Foundation
+
+protocol AlertPresenterProtocol {
+    func show(alertModel: AlertModel)
+}
