+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  QuestionFactoryProtocol.swift
+//  MovieQuiz
+//
+//  Created by Александра Коснырева on 07.09.2023.
+//
+
+import Foundation
+
+protocol QuestionFactoryProtocol {
+    func requestNextQuestion() 
+}
